backend/requirements.txt
flask
flask-cors
flask-socketio
pymongo
python-dotenv
bcrypt
pyjwt
flask-bcrypt
dnspython
cloudinary
eventlet
